Move temp file class into its own file
