+++ omeroweb/webgateway/views.py
@@ -76,8 +76,8 @@ import glob
 from omeroweb.webgateway.webgateway_cache import (
     webgateway_cache,
     CacheBase,
-    webgateway_tempfile,
 )
+from omeroweb.webgateway.webgateway_tempfile import webgateway_tempfile
 
 import logging
 import os
